move function to tauri_utils
