+++ tooling/bundler/src/lib.rs
@@ -21,4 +21,3 @@ pub mod bundle;
 mod error;
 pub use bundle::*;
 pub use error::{Error, Result};
-mod util;
