--- tooling/bundler/src/util.rs
@@ -1,9 +1,0 @@
-use std::path::{Path, PathBuf};
-
-pub fn display_path<P: AsRef<Path>>(p: P) -> String {
-  p.as_ref()
-    .components()
-    .collect::<PathBuf>()
-    .to_string_lossy()
-    .to_string()
-}
