+++ tooling/cli/src/interface/rust.rs
@@ -36,8 +36,8 @@ use super::{AppSettings, ExitReason, Interface};
 use crate::helpers::{
   app_paths::{app_dir, tauri_dir},
   config::{nsis_settings, reload as reload_config, wix_settings, Config},
-  display_path,
 };
+use tauri_utils::display_path;
 
 mod cargo_config;
 mod desktop;
